+++ packages/inject/src/main.ts
@@ -1,6 +1,8 @@
 export {
   type Inject,
   Injector,
+  type Register,
   type SafeInject,
   type Type,
+  type Unregister,
 } from "./injector.ts";